test: filters - refactor to use block method
