+++ lib/primitives/block.js
@@ -21,6 +21,7 @@ const Network = require('../protocol/network');
 const util = require('../utils/util');
 const {encoding} = bio;
 const {inspectSymbol} = require('../utils');
+const GCSFilter = require('golomb');
 
 /**
  * Block
@@ -838,6 +839,40 @@ class Block extends AbstractBlock {
 
   static isBlock(obj) {
     return obj instanceof Block;
+  }
+
+  toFilter(view) {
+    const hash = this.hash();
+    const key = hash.slice(0, 16);
+    const items = new BufferSet();
+
+    for (let i = 0; i < this.txs.length; i++) {
+      const tx = this.txs[i];
+
+      for (const output of tx.outputs) {
+        if (output.script.length === 0)
+          continue;
+
+        // In order to allow the filters to later be committed
+        // to within an OP_RETURN output, we ignore all
+        // OP_RETURNs to avoid a circular dependency.
+        if (output.script.raw[0] === 106)
+          continue;
+
+        items.add(output.script.raw);
+      }
+    }
+
+    for (const [, coins] of view.map) {
+      for (const [, coin] of coins.outputs) {
+        if (coin.output.script.length === 0)
+          continue;
+
+        items.add(coin.output.script.raw);
+      }
+    }
+
+    return GCSFilter.fromItems(19, key, items);
   }
 }
